Move prepare_cloned_fields() to utilities.querydict
